feat: Images (#21)
This PR is adding the saving of images for Spaces and Boxes.
The feature adds dependency to a Azure Storage Account and Azure Functions. The functions runs as managed functions within the context of the Azure Static Web App that is hosting everything.

This PR also brings deletion of boxes and some bug fixes along with updated documentation.